Add convert_to_image_url(), which takes a regular imgur page url and returns the direct URL of the image on that page.
This allows repost check even if the user pastes in an imgur page URL by accident.

One thing to note: the image URL preview functionality still warns that the URL is not a valid image URL. Whether or not we change this part too is a design decision that we leave for later.